move convertPremiumIndex to convert.go
